+++ sdk/nodejs/tests/runtime/langhost/cases/061.provider_in_parent_invokes/index.js
@@ -27,21 +27,7 @@ let args = {
     urn: "some-urn",
 };
 
-if (semver.lt(process.version, "12.11.0")) {
-	// These tests hang on runtimes later than 12.10.x due to their use of deasync.
-
-	let result1 = pulumi.runtime.invoke("test:index:echo", args, { parent, async: false });
-	for (const key in args) {
-		assert.deepEqual(result1[key], args[key]);
-	}
-
-	let result2 = pulumi.runtime.invoke("test:index:echo", args, { parent, async: false });
-	result2.then((v) => {
-		assert.deepEqual(v, args);
-	});
-}
-
-let result3 = pulumi.runtime.invoke("test:index:echo", args, { parent });
+let result3 = pulumi.runtime.invoke("test:index:echo", args, { parent, async: true });
 result3.then((v) => {
     assert.deepEqual(v, args);
 });